requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
CacheControl==0.12.6
celery==5.0.5
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
colorclass==2.2.0
contextlib2==0.6.0.post1
distlib==0.3.1
distro==1.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-mysql==3.9.0
djangorestframework==3.12.2
docopt==0.6.2
html5lib==1.1
idna==2.10
ipaddr==2.2.0
kombu==5.0.2
lockfile==0.12.2
Markdown==3.3.3
msgpack==1.0.0
mysqlclient==2.0.1
packaging==20.4
pep517==0.9.1
pip-upgrader==1.4.15
progress==1.5
prompt-toolkit==3.0.8
pyparsing==2.4.7
pytoml==0.1.21
pytz==2020.4
redis==3.5.3
requests==2.25.0
retrying==1.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
terminaltables==3.1.0
toml==0.10.2
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
